+++ app/src/main/AndroidManifest.xml
@@ -17,13 +17,22 @@
             android:windowSoftInputMode="adjustResize"
             android:exported="false" />
         <activity
-            android:name=".activities.all.HomePageActivity"
+            android:name=".activities.TriviaActivity"
             android:exported="false" />
         <activity
-            android:name=".activities.all.SettingsPageActivity"
+            android:name=".activities.SettingsPageActivity"
             android:exported="false" />
         <activity
-            android:name=".activities.all.ProfilePageActivity"
+            android:name=".activities.ProfilePageActivity"
+            android:exported="false" />
+        <activity
+            android:name=".activities.authentication.LogInActivity"
+            android:exported="false" />
+        <activity
+            android:name=".activities.authentication.SignUpActivity"
+            android:exported="false" />
+        <activity
+            android:name=".activities.authentication.VerifyAccountActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
