+++ app/src/main/java/com/mca/mindmelter/activities/ProfilePageActivity.java
@@ -1,0 +1,26 @@
+package com.mca.mindmelter.activities;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+
+import com.mca.mindmelter.R;
+
+public class ProfilePageActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_profile_page);
+
+        setUpSettingsButton();
+    }
+
+    public void setUpSettingsButton() {
+        findViewById(R.id.mainActivitySettingsImageView).setOnClickListener(view -> {
+            Intent goToSettingsPageIntent = new Intent(ProfilePageActivity.this, SettingsPageActivity.class);
+            startActivity(goToSettingsPageIntent);
+        });
+    }
+}
