--- app/src/main/java/com/mca/mindmelter/activities/all/SettingsPageActivity.java
@@ -1,51 +1,0 @@
-package com.mca.mindmelter.activities.all;
-
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.app.AppCompatDelegate;
-
-import android.content.Context;
-import android.content.SharedPreferences;
-import android.os.Bundle;
-import android.widget.Switch;
-
-import com.mca.mindmelter.R;
-
-public class SettingsPageActivity extends AppCompatActivity {
-    Switch switcher;
-    boolean nightMode;
-    SharedPreferences sharedPreferences;
-    SharedPreferences.Editor editor;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_settings_page);
-
-        switcher = findViewById(R.id.modeSwitch);
-
-        setUpThemeSwitch();
-    }
-
-    public void setUpThemeSwitch(){
-        // we use sharedPreferences to save mode if you exit the app and go back
-        sharedPreferences = getSharedPreferences("MODE", Context.MODE_PRIVATE);
-        nightMode = sharedPreferences.getBoolean("night", false); //light mode is default
-
-        if(nightMode){
-            switcher.setChecked(true);
-            AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES);
-        }
-        switcher.setOnClickListener(view ->{
-            if (nightMode){
-                AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO);
-                editor = sharedPreferences.edit();
-                editor.putBoolean("night", false);
-            }else{
-                AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES);
-                editor = sharedPreferences.edit();
-                editor.putBoolean("night", true);
-            }
-            editor.apply();
-        });
-    }
-}
