+++ app/src/main/res/layout/activity_trivia.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".activities.TriviaActivity"
+    style="@style/SCREEN">
+
+    <Button
+        android:id="@+id/homePageLearnMoreButton"
+        android:layout_width="225dp"
+        android:layout_height="48dp"
+        android:layout_marginBottom="264dp"
+        android:text="@string/learn_more"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        style="@style/BUTTON"/>
+
+    <TextView
+        android:id="@+id/homePageTextView"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="60dp"
+        android:text="@string/on_this_day_text"
+        android:textSize="30sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.498"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        android:textStyle="bold"
+        style="@style/TEXT"/>
+
+    <TextView
+        android:id="@+id/homePageTriviaTextView"
+        android:layout_width="280dp"
+        android:layout_height="268dp"
+        android:layout_marginTop="24dp"
+        android:text="@string/home_page_trivia"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.496"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/homePageTextView"
+        tools:ignore="TextSizeCheck"
+        style="@style/TEXT"/>
+
+</androidx.constraintlayout.widget.ConstraintLayout>
